+++ demo/server/src/app_state.rs
@@ -1,6 +1,6 @@
 use std::collections::HashSet;
 use std::sync::Arc;
-use tokio::sync::Mutex;
+use tokio::sync::RwLock;
 
 use crate::connection::{Connection, SocketError};
 
@@ -8,8 +8,8 @@ use std::sync::OnceLock;
 
 #[derive(Clone)]
 pub struct AppState {
-    connections: Arc<Mutex<HashSet<Connection>>>,
-    messages: Arc<Mutex<Vec<String>>>,
+    connections: Arc<RwLock<HashSet<Connection>>>,
+    messages: Arc<RwLock<Vec<String>>>,
 }
 
 impl AppState {
@@ -20,28 +20,28 @@ impl AppState {
 
     fn new() -> AppState {
         AppState {
-            connections: Arc::new(Mutex::new(HashSet::new())),
-            messages: Arc::new(Mutex::new(Vec::new())),
+            connections: Arc::new(RwLock::new(HashSet::new())),
+            messages: Arc::new(RwLock::new(Vec::new())),
         }
     }
 
     pub async fn add_connection(&self, socket: &Connection) {
-        let mut connections = self.connections.lock().await;
+        let mut connections = self.connections.write().await;
         connections.insert(socket.clone());
     }
 
     pub async fn remove_connection(&self, user_name: &Connection) {
-        let _ = self.connections.lock().await.remove(user_name);
+        let _ = self.connections.write().await.remove(user_name);
     }
 
     async fn add_message(&self, line: &str) {
-        self.messages.lock().await.push(line.to_string());
+        self.messages.write().await.push(line.to_string());
     }
 
     pub async fn message_from(&self, message: String) -> Result<(), SocketError> {
         self.add_message(&message).await;
 
-        for connection in self.connections.lock().await.iter() {
+        for connection in self.connections.read().await.iter() {
             let _ = connection.send(&message).await;
         }
 
@@ -49,7 +49,7 @@ impl AppState {
     }
 
     pub async fn send_all_prev_messages(&self, socket: &Connection) -> Result<(), SocketError> {
-        let all_messages = self.messages.lock().await;
+        let all_messages = self.messages.read().await;
 
         for message in all_messages.iter() {
             socket.send(message).await?;